+++ src/components/Icon/iconRules.ts
@@ -1,4 +1,5 @@
 import fontAwesomeIcons from './fontAwesomeIconRules'
+import { disabledStyles } from '../../styles/customCSS'
 
 const sizes = new Map([
   ['mini', 0.4],
@@ -38,7 +39,7 @@ const getBorderedStyles = (circular, borderColor, color) => ({
 })
 
 const iconRules = {
-  root: ({ props: { color, kind, name, size, bordered, circular }, variables: v }) => {
+  root: ({ props: { color, disabled, kind, name, size, bordered, circular }, variables: v }) => {
     const { fontFamily, content } = getIcon(kind, name)
     const iconColor = color || v.color
 
@@ -69,6 +70,8 @@ const iconRules = {
       },
 
       ...((bordered || circular) && getBorderedStyles(circular, v.borderColor, iconColor)),
+
+      ...(disabled && disabledStyles),
     }
   },
 }